# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Authentication API",
      "description": "API for user authentication and profile management",
      "version": "1.0.0"
    },
    "host": "localhost:3000",
    "schemes": ["http"],
    "paths": {
      "/api/auth/register": {
        "post": {
          "summary": "Register a new user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User registered successfully"
            }
          }
        }
      },
      "/api/auth/login": {
        "post": {
          "summary": "Login a user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Login"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User logged in successfully"
            }
          }
        }
      },
      "/api/profile/me": {
        "get": {
          "summary": "Get current user's profile",
          "security": [{ "Bearer": [] }],
          "responses": {
            "200": {
              "description": "User profile retrieved successfully"
            }
          }
        },
        "put": {
          "summary": "Update current user's profile",
          "security": [{ "Bearer": [] }],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User profile updated successfully"
            }
          }
        }
      },
      "/api/profile/public": {
        "get": {
          "summary": "Get all public profiles",
          "responses": {
            "200": {
              "description": "Public profiles retrieved successfully"
            }
          }
        }
      },
      "/api/profile/all": {
        "get": {
          "summary": "Get all profiles (admin only)",
          "security": [{ "Bearer": [] }],
          "responses": {
            "200": {
              "description": "All profiles retrieved successfully"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "bio": { "type": "string" },
          "phone": { "type": "string" },
          "photo": { "type": "string" },
          "isPublic": { "type": "boolean" }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      }
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
  